+++ mobile/android/geckoview/src/androidTest/assets/web_extensions/optional-permission-request/request-permission.js
@@ -3,6 +3,7 @@ window.onload = () => {
     const perms = {
       permissions: ["activeTab"],
       origins: ["*://example.com/*"],
+      data_collection: ["healthInfo"],
     };
     const response = await browser.permissions.request(perms);
     browser.runtime.sendNativeMessage("browser", `${response}`);