+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/WebExtensionTest.kt
@@ -205,6 +205,7 @@ class WebExtensionTest : BaseSessionTest() {
                 "xpinstall.signatures.required" to false,
                 "extensions.install.requireBuiltInCerts" to false,
                 "extensions.update.requireBuiltInCerts" to false,
+                "extensions.dataCollectionPermissions.enabled" to true,
             ),
         )
 
@@ -222,6 +223,7 @@ class WebExtensionTest : BaseSessionTest() {
 
         var grantedOptionalPermissions = extension.metaData.grantedOptionalPermissions
         var grantedOptionalOrigins = extension.metaData.grantedOptionalOrigins
+        var grantedOptionalDataCollectionPermissions = extension.metaData.grantedOptionalDataCollectionPermissions
 
         assertThat(
             "grantedOptionalPermissions must be 0.",
@@ -229,6 +231,11 @@ class WebExtensionTest : BaseSessionTest() {
             equalTo(0),
         )
         assertThat("grantedOptionalOrigins must be 0.", grantedOptionalOrigins.size, equalTo(0))
+        assertThat(
+            "grantedOptionalDataCollectionPermissions must be 0.",
+            grantedOptionalDataCollectionPermissions.size,
+            equalTo(0),
+        )
 
         // click triggers permissions.request
         mainSession.synthesizeTap(50, 50)
@@ -238,6 +245,7 @@ class WebExtensionTest : BaseSessionTest() {
                 extension: WebExtension,
                 permissions: Array<String>,
                 origins: Array<String>,
+                dataCollectionPermissions: Array<String>,
             ): GeckoResult<AllowOrDeny> {
                 return GeckoResult.allow()
             }
@@ -270,6 +278,7 @@ class WebExtensionTest : BaseSessionTest() {
 
         grantedOptionalPermissions = updatedExtension.metaData.grantedOptionalPermissions
         grantedOptionalOrigins = updatedExtension.metaData.grantedOptionalOrigins
+        grantedOptionalDataCollectionPermissions = updatedExtension.metaData.grantedOptionalDataCollectionPermissions
 
         assertThat(
             "grantedOptionalPermissions must be 1.",
@@ -282,6 +291,12 @@ class WebExtensionTest : BaseSessionTest() {
             grantedOptionalOrigins.first(),
             equalTo("*://example.com/*"),
         )
+        assertThat(
+            "grantedOptionalDataCollectionPermissions must be 1.",
+            grantedOptionalDataCollectionPermissions.size,
+            equalTo(1),
+        )
+        assertArrayEquals(arrayOf("healthInfo"), grantedOptionalDataCollectionPermissions)
 
         // geolocation is part of the manifest but not requested/granted.
         assertFalse(grantedOptionalPermissions.contains("geolocation"))
@@ -298,6 +313,12 @@ class WebExtensionTest : BaseSessionTest() {
 
     @Test
     fun updateOptionalPermissions() {
+        sessionRule.setPrefsUntilTestEnd(
+            mapOf(
+                "extensions.dataCollectionPermissions.enabled" to true,
+            ),
+        )
+
         var extension = sessionRule.waitForResult(
             controller.ensureBuiltIn(
                 "resource://android/assets/web_extensions/optional-permission-request/",
@@ -309,6 +330,7 @@ class WebExtensionTest : BaseSessionTest() {
 
         var grantedOptionalPermissions = extension.metaData.grantedOptionalPermissions
         var grantedOptionalOrigins = extension.metaData.grantedOptionalOrigins
+        var grantedOptionalDataCollectionPermissions = extension.metaData.grantedOptionalDataCollectionPermissions
 
         // Without adding any optional permissions.
         assertThat(
@@ -317,6 +339,11 @@ class WebExtensionTest : BaseSessionTest() {
             equalTo(0),
         )
         assertThat("grantedOptionalOrigins must be 0.", grantedOptionalOrigins.size, equalTo(0))
+        assertThat(
+            "grantedOptionalDataCollectionPermissions must be 0.",
+            grantedOptionalDataCollectionPermissions.size,
+            equalTo(0),
+        )
 
         // Only adding an origin permission.
         extension = sessionRule.waitForResult(
@@ -330,6 +357,7 @@ class WebExtensionTest : BaseSessionTest() {
 
         grantedOptionalPermissions = extension.metaData.grantedOptionalPermissions
         grantedOptionalOrigins = extension.metaData.grantedOptionalOrigins
+        grantedOptionalDataCollectionPermissions = extension.metaData.grantedOptionalDataCollectionPermissions
 
         assertThat(
             "grantedOptionalPermissions must be 0.",
@@ -341,6 +369,11 @@ class WebExtensionTest : BaseSessionTest() {
             "grantedOptionalOrigins must be *://example.com/*.",
             grantedOptionalOrigins.first(),
             equalTo("*://example.com/*"),
+        )
+        assertThat(
+            "grantedOptionalDataCollectionPermissions must be 0.",
+            grantedOptionalDataCollectionPermissions.size,
+            equalTo(0),
         )
 
         // Adding "nothing" to verify that nothing gets changed.
@@ -359,6 +392,11 @@ class WebExtensionTest : BaseSessionTest() {
             grantedOptionalOrigins.first(),
             equalTo("*://example.com/*"),
         )
+        assertThat(
+            "grantedOptionalDataCollectionPermissions must be 0.",
+            grantedOptionalDataCollectionPermissions.size,
+            equalTo(0),
+        )
 
         // Adding an optional permission.
         extension = sessionRule.waitForResult(
@@ -372,6 +410,7 @@ class WebExtensionTest : BaseSessionTest() {
 
         grantedOptionalPermissions = extension.metaData.grantedOptionalPermissions
         grantedOptionalOrigins = extension.metaData.grantedOptionalOrigins
+        grantedOptionalDataCollectionPermissions = extension.metaData.grantedOptionalDataCollectionPermissions
 
         // Both optional and origin permissions must be granted.
         assertThat(
@@ -390,6 +429,11 @@ class WebExtensionTest : BaseSessionTest() {
             grantedOptionalOrigins.first(),
             equalTo("*://example.com/*"),
         )
+        assertThat(
+            "grantedOptionalDataCollectionPermissions must be 0.",
+            grantedOptionalDataCollectionPermissions.size,
+            equalTo(0),
+        )
 
         // Removing "nothing" to verify that nothing gets changed.
         extension = sessionRule.waitForResult(
@@ -407,6 +451,11 @@ class WebExtensionTest : BaseSessionTest() {
             grantedOptionalOrigins.first(),
             equalTo("*://example.com/*"),
         )
+        assertThat(
+            "grantedOptionalDataCollectionPermissions must be 0.",
+            grantedOptionalDataCollectionPermissions.size,
+            equalTo(0),
+        )
 
         // Remove an activeTab optional permission.
         extension = sessionRule.waitForResult(
@@ -420,6 +469,7 @@ class WebExtensionTest : BaseSessionTest() {
 
         grantedOptionalPermissions = extension.metaData.grantedOptionalPermissions
         grantedOptionalOrigins = extension.metaData.grantedOptionalOrigins
+        grantedOptionalDataCollectionPermissions = extension.metaData.grantedOptionalDataCollectionPermissions
 
         assertThat(
             "grantedOptionalPermissions must be 0.",
@@ -431,6 +481,11 @@ class WebExtensionTest : BaseSessionTest() {
             "grantedOptionalOrigins must be *://example.com/*.",
             grantedOptionalOrigins.first(),
             equalTo("*://example.com/*"),
+        )
+        assertThat(
+            "grantedOptionalDataCollectionPermissions must be 0.",
+            grantedOptionalDataCollectionPermissions.size,
+            equalTo(0),
         )
 
         // Remove an `*://example.com/*` origin permission.
@@ -445,6 +500,7 @@ class WebExtensionTest : BaseSessionTest() {
 
         grantedOptionalPermissions = extension.metaData.grantedOptionalPermissions
         grantedOptionalOrigins = extension.metaData.grantedOptionalOrigins
+        grantedOptionalDataCollectionPermissions = extension.metaData.grantedOptionalDataCollectionPermissions
 
         assertThat(
             "grantedOptionalPermissions must be 0.",
@@ -452,6 +508,33 @@ class WebExtensionTest : BaseSessionTest() {
             equalTo(0),
         )
         assertThat("grantedOptionalOrigins must be 0.", grantedOptionalOrigins.size, equalTo(0))
+        assertThat(
+            "grantedOptionalDataCollectionPermissions must be 0.",
+            grantedOptionalDataCollectionPermissions.size,
+            equalTo(0),
+        )
+
+        // Adding an optional data collection permission.
+        extension = sessionRule.waitForResult(
+            controller.addOptionalPermissions(
+                extension.id,
+                arrayOf(),
+                arrayOf(),
+                arrayOf("healthInfo"),
+            ),
+        )
+
+        grantedOptionalPermissions = extension.metaData.grantedOptionalPermissions
+        grantedOptionalOrigins = extension.metaData.grantedOptionalOrigins
+        grantedOptionalDataCollectionPermissions = extension.metaData.grantedOptionalDataCollectionPermissions
+
+        assertThat(
+            "grantedOptionalPermissions must be 0.",
+            grantedOptionalPermissions.size,
+            equalTo(0),
+        )
+        assertThat("grantedOptionalOrigins must be 0.", grantedOptionalOrigins.size, equalTo(0))
+        assertArrayEquals(arrayOf("healthInfo"), grantedOptionalDataCollectionPermissions)
 
         // Missing origins from the manifest.
         // Must throw!
@@ -526,6 +609,7 @@ class WebExtensionTest : BaseSessionTest() {
         sessionRule.setPrefsUntilTestEnd(
             mapOf(
                 "extensions.originControls.grantByDefault" to false,
+                "extensions.dataCollectionPermissions.enabled" to true,
             ),
         )
 
@@ -567,6 +651,12 @@ class WebExtensionTest : BaseSessionTest() {
 
     @Test
     fun onOptionalPermissionsChanged() {
+        sessionRule.setPrefsUntilTestEnd(
+            mapOf(
+                "extensions.dataCollectionPermissions.enabled" to true,
+            ),
+        )
+
         var extension = sessionRule.waitForResult(
             controller.ensureBuiltIn(
                 "resource://android/assets/web_extensions/optional-permission-request/",
@@ -578,6 +668,7 @@ class WebExtensionTest : BaseSessionTest() {
 
         var grantedOptionalPermissions = extension.metaData.grantedOptionalPermissions
         var grantedOptionalOrigins = extension.metaData.grantedOptionalOrigins
+        var grantedOptionalDataCollectionPermissions = extension.metaData.grantedOptionalDataCollectionPermissions
 
         assertThat(
             "grantedOptionalPermissions must be 0.",
@@ -585,12 +676,20 @@ class WebExtensionTest : BaseSessionTest() {
             equalTo(0),
         )
         assertThat("grantedOptionalOrigins must be 0.", grantedOptionalOrigins.size, equalTo(0))
+        assertThat(
+            "grantedOptionalDataCollectionPermissions must be 0.",
+            grantedOptionalDataCollectionPermissions.size,
+            equalTo(0),
+        )
 
         sessionRule.delegateDuringNextWait(object : WebExtensionController.AddonManagerDelegate {
             @AssertCalled(count = 1)
             override fun onOptionalPermissionsChanged(updatedExtension: WebExtension) {
                 grantedOptionalPermissions = updatedExtension.metaData.grantedOptionalPermissions
                 grantedOptionalOrigins = updatedExtension.metaData.grantedOptionalOrigins
+                grantedOptionalDataCollectionPermissions =
+                    updatedExtension.metaData.grantedOptionalDataCollectionPermissions
+
                 assertNull(updatedExtension)
                 assertArrayEquals(
                     "grantedOptionalPermissions must be [activeTab, geolocation].",
@@ -602,6 +701,7 @@ class WebExtensionTest : BaseSessionTest() {
                     arrayOf("*://example.com/*"),
                     grantedOptionalOrigins,
                 )
+                assertArrayEquals(arrayOf("healthInfo"), grantedOptionalDataCollectionPermissions)
             }
         })
 
@@ -610,7 +710,7 @@ class WebExtensionTest : BaseSessionTest() {
                 extension.id,
                 arrayOf("activeTab", "geolocation"),
                 arrayOf("*://example.com/*"),
-                arrayOf(),
+                arrayOf("healthInfo"),
             ),
         )
         sessionRule.waitForResult(controller.uninstall(extension))
@@ -3243,7 +3343,12 @@ class WebExtensionTest : BaseSessionTest() {
 
         sessionRule.delegateUntilTestEnd(object : WebExtensionController.PromptDelegate {
             @AssertCalled(count = 2)
-            override fun onOptionalPrompt(extension: WebExtension, permissions: Array<String>, origins: Array<String>): GeckoResult<AllowOrDeny> {
+            override fun onOptionalPrompt(
+                extension: WebExtension,
+                permissions: Array<String>,
+                origins: Array<String>,
+                dataCollectionPermissions: Array<String>,
+            ): GeckoResult<AllowOrDeny> {
                 val expected = arrayOf("geolocation")
                 assertThat("Permissions should match the requested permissions", permissions, equalTo(expected))
                 assertThat("Origins should match the requested origins", origins, equalTo(arrayOf("*://example.com/*")))